create rubyScript 'tree_Codewars_v_1_0_1'

+++ Ruby/creating_a_unique_folder/folder_name_time_now.rb
@@ -4,7 +4,7 @@ folder_name = ARGV[0]
 def create_uniq_folder(folder_name = nil)
     folder_name  ||= 'folder_name'
    if Dir.exist?("./#{folder_name}")
-        suffix = Time.now.to_f.to_s.gsub('.','')
+        suffix = Time.now.to_i.to_s.gsub('.','')
         loop do
            folder_path = "#{folder_name}#{suffix}"
             unless Dir.exist? folder_path
